refactor: Player 객체 수정
